requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
email-validator==2.1.1
SQLAlchemy==2.0.36
